requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
sqlalchemy==2.0.35
psycopg[binary]
python-dotenv==1.0.1
pydantic-settings==2.2.1
alembic==1.13.1
reportlab==4.2.0
openpyxl==3.1.2 
xlsxwriter==3.2.0


httpx==0.27.0
aiocache==0.12.2
redis==5.0.3 
aioredis==2.0.1 


cryptography==42.0.5
pyjwt==2.8.0
passlib[bcrypt]==1.7.4
fastapi-limiter==0.1.6
sentry-sdk[fastapi]==1.42.0


gunicorn==22.0.0